Added Welcome, Login, and Registration HTML Pages
Welcome.html needs functional buttons that point to login component  and registration component; login and registration used the specified Angular Data Bindings that we used for our TS file so I'm think we don't need to change anything with them. I also added scripts and css links in the index.html file which may need to be changed. But I did do "ng serve" and it successfully compiled and displayed on Google Chrome. If you want to see it for yourself, do a quick pull from my "JacenF" branch.